refactor(proto): split pob2.proto into domain-specific files
Split the monolithic 502-line pob2.proto into 6 domain-specific files
for better maintainability and organization:

- common.proto: Position, StatSnapshot, StatDelta, CharacterClass
- tree.proto: PassiveNode, PassiveTree, NodeGroup, TreeConstants, NodeType
- items.proto: Item, Socket, *Data, ItemRarity, ItemType, ItemSlot
- skills.proto: SkillGroup, GemInstance, Gem, GemLevel, GemType
- build.proto: Build, BuildConfig
- leveling.proto: LevelingPath, Checkpoint, LevelingStep, triggers/actions

Updated all TypeScript imports throughout the codebase to use the
domain-specific proto files instead of the combined pob2_pb.

Also:
- Updated proto:generate script to process all 6 files
- Added generated *_pb.ts files to .gitignore
- Removed old pob2.proto file

Closes #132
