+++ src/proto/items.proto
@@ -1,0 +1,184 @@
+// items.proto
+// Path of Building 2 - Web Edition
+// Item types and related data
+
+syntax = "proto3";
+
+package pob2;
+
+// ============================================================================
+// ITEMS
+// ============================================================================
+
+message Item {
+  string id = 1;  // REQUIRED - unique item identifier
+
+  // Identity
+  optional string name = 2;
+  optional string base_name = 3;
+  optional string type_line = 4;
+  optional ItemRarity rarity = 5;
+  optional ItemType item_type = 6;
+
+  // Properties
+  optional int32 item_level = 10;
+  optional int32 quality = 11;
+  optional bool corrupted = 12;
+  optional bool mirrored = 13;
+  optional bool fractured = 14;
+
+  // Requirements
+  optional int32 required_level = 20;
+  optional int32 required_str = 21;
+  optional int32 required_dex = 22;
+  optional int32 required_int = 23;
+
+  // Sockets & Runes (PoE2)
+  repeated Socket sockets = 30;
+  repeated string runes = 31;
+
+  // Modifiers
+  repeated string implicit_mods = 40;
+  repeated string explicit_mods = 41;
+  repeated string enchant_mods = 42;
+  repeated string rune_mods = 43;
+  repeated string crafted_mods = 44;
+
+  // Weapon data (if applicable)
+  optional WeaponData weapon_data = 50;
+
+  // Armour data (if applicable)
+  optional ArmourData armour_data = 51;
+
+  // Flask data (if applicable)
+  optional FlaskData flask_data = 52;
+
+  // Jewel data (if applicable)
+  optional JewelData jewel_data = 53;
+}
+
+message Socket {
+  optional string color = 1;  // R, G, B, W, A (abyss)
+  optional int32 group = 2;   // Link group
+}
+
+message WeaponData {
+  optional int32 physical_min = 1;
+  optional int32 physical_max = 2;
+  optional int32 fire_min = 3;
+  optional int32 fire_max = 4;
+  optional int32 cold_min = 5;
+  optional int32 cold_max = 6;
+  optional int32 lightning_min = 7;
+  optional int32 lightning_max = 8;
+  optional int32 chaos_min = 9;
+  optional int32 chaos_max = 10;
+  optional float attack_speed = 11;
+  optional float crit_chance = 12;  // Percentage * 100 (e.g., 650 = 6.50%)
+  optional int32 range = 13;
+}
+
+message ArmourData {
+  optional int32 armour = 1;
+  optional int32 evasion = 2;
+  optional int32 energy_shield = 3;
+  optional int32 ward = 4;
+  optional int32 block = 5;  // Percentage * 100
+}
+
+message FlaskData {
+  optional int32 life_recovery = 1;
+  optional int32 mana_recovery = 2;
+  optional int32 duration = 3;  // Milliseconds
+  optional int32 charges_used = 4;
+  optional int32 charges_max = 5;
+}
+
+message JewelData {
+  optional int32 radius = 1;
+  optional string limit = 2;  // Limit text
+  optional bool is_cluster = 3;
+  optional int32 cluster_size = 4;  // Small, Medium, Large
+}
+
+// ============================================================================
+// ITEM ENUMS
+// ============================================================================
+
+enum ItemRarity {
+  ITEM_RARITY_UNKNOWN = 0;
+  RARITY_NORMAL = 1;
+  RARITY_MAGIC = 2;
+  RARITY_RARE = 3;
+  RARITY_UNIQUE = 4;
+}
+
+enum ItemType {
+  ITEM_TYPE_UNKNOWN = 0;
+
+  // Weapons
+  ONE_HAND_SWORD = 1;
+  TWO_HAND_SWORD = 2;
+  ONE_HAND_AXE = 3;
+  TWO_HAND_AXE = 4;
+  ONE_HAND_MACE = 5;
+  TWO_HAND_MACE = 6;
+  BOW = 7;
+  CROSSBOW = 8;
+  STAFF = 9;
+  WAND = 10;
+  DAGGER = 11;
+  CLAW = 12;
+  SCEPTRE = 13;
+  QUARTERSTAFF = 14;  // PoE2
+  SPEAR = 15;         // PoE2
+  FLAIL = 16;         // PoE2
+
+  // Armour
+  HELMET = 20;
+  BODY_ARMOUR = 21;
+  GLOVES = 22;
+  BOOTS = 23;
+  SHIELD = 24;
+  FOCUS = 25;  // PoE2 off-hand
+
+  // Accessories
+  AMULET = 30;
+  RING = 31;
+  BELT = 32;
+
+  // Other
+  FLASK = 40;
+  JEWEL = 41;
+  CHARM = 42;  // PoE2
+  QUIVER = 43;
+}
+
+enum ItemSlot {
+  ITEM_SLOT_UNKNOWN = 0;
+
+  SLOT_WEAPON_1 = 1;
+  SLOT_WEAPON_2 = 2;  // Off-hand / Shield / Focus
+  SLOT_WEAPON_1_SWAP = 3;
+  SLOT_WEAPON_2_SWAP = 4;
+
+  SLOT_HELMET = 10;
+  SLOT_BODY_ARMOUR = 11;
+  SLOT_GLOVES = 12;
+  SLOT_BOOTS = 13;
+
+  SLOT_AMULET = 20;
+  SLOT_RING_1 = 21;
+  SLOT_RING_2 = 22;
+  SLOT_BELT = 23;
+
+  SLOT_FLASK_1 = 30;
+  SLOT_FLASK_2 = 31;
+  SLOT_FLASK_3 = 32;
+  SLOT_FLASK_4 = 33;
+  SLOT_FLASK_5 = 34;
+
+  SLOT_JEWEL_1 = 40;
+  SLOT_JEWEL_2 = 41;
+  // ... more jewel slots as needed
+}
